requirements.txt
# Runtime dependencies
Flask==1.1.1
Flask-API==1.1
flask-restplus==0.13.0
Flask-SQLAlchemy==2.4.1
psycopg2-binary==2.8.3

# runtime
gunicorn==19.9.0
honcho==1.0.1

# Database

# Test Driven Development
nose==1.3.7
pinocchio==0.4.2
factory-boy==2.12.0
pylint>=2.4.1

# Behavior Driven Development
behave==1.2.6
selenium==3.141.0
compare==0.2b0
requests==2.22.0

# Code coverage
coverage==4.5.4
codecov==2.0.15